// src/app/modules/einvoice/models/upload-einv-result.model.ts
export interface UploadEinvResult {
    documentNumber: string;
    documentDt: string;
    buyerGstin: string
    buyerName: string;
    buyerAddressLine1: string;
    buyerAddressLine2: string;
    buyerAddressLine3: string;
    buyerLocation: string;
    buyerPin: string;
    buyerStateCode: string;
    buyerPlaceOfSupply:string;
    markAsB2C: string;
}